Fix review findings: wrap dbt/dlt write-through errors, distinguish Metabase auth failure
Second narrower review pass (scoped to the metabase state-cache, the 8
newly-wired dbt call sites, and the new except Exception handler) found
two more real issues, both verified before fixing:

1. _set_dbt_telemetry() and _set_dlt_telemetry() raised unwrapped
   OSError/tomlkit.exceptions.TOMLKitError instead of click.ClickException,
   breaking the --all graceful-degradation contract every other provider
   follows (a read-only ~/.dango, or a hand-corrupted .dlt/config.toml,
   would crash the whole `--all` command with a raw traceback instead of
   skipping that provider and continuing). Both now wrap their filesystem/
   tomlkit operations and convert to click.ClickException, matching the
   contract set_metabase_telemetry() already uses.

2. A Metabase login 401/403 (stale admin password in .dango/metabase.yml)
   was mislabeled "Could not reach Metabase — is it running?" since
   requests.HTTPError is a RequestException subclass. Metabase is running
   fine in this case; the credentials are just wrong. Now checked
   explicitly on the login response status code before raise_for_status(),
   with a distinct "check admin credentials in .dango/metabase.yml"
   message.

7 new regression tests covering both fixes. Also split
tests/unit/test_telemetry_unified.py (which had grown to 568 lines,
tripping the 500-line file-size-check pre-commit hook) into two files
along its existing class boundaries: dbt/dlt + top-level CLI tests stay
in test_telemetry_unified.py, all Metabase-specific tests move to the new
test_telemetry_unified_metabase.py. No exemption added — this is new code
from this PR, not legacy debt, so splitting is the correct fix rather than
reaching for docs/file-exemptions.yml.
